dao exec: extract exec logic into handler for reusability
